% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_dirlist.R
\name{download_gsefiles}
\alias{download_gsefiles}
\title{Download files from Entrez GEO repository}
\usage{
download_gsefiles(gsefiles, dest = ".", verbose = FALSE)
}
\arguments{
\item{gsefiles}{GEO file names, starting with Accession number, a character vector.}

\item{dest}{Path to local folder where downloaded files will be stored, a character string. Defaults to current directory.}

\item{verbose}{Defaults to FALSE.}

\item{sleep}{sleep between queries in seconds, a double.}
}
\description{
Downloads supplementary or MINiML formatted family files from Entrez GEO repository.
}
\examples{
\dontrun{
c("GSE100206_family.xml.tgz",
"GSE100206_exoNormal_circRNA_RPM.txt.gz") \%>\%
download_gsefiles()
}

}
