% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrez_docsums.R
\name{extract_docsums}
\alias{extract_docsums}
\title{Extract GEO DocSums into tibble}
\usage{
extract_docsums(xmldoc)
}
\arguments{
\item{xmldoc}{A GEO query result contents, list of document summaries of class "xml_document" "xml_node"}
}
\value{
A tibble of GEO document summaries
}
\description{
Extracts entrez esummary API results from XML into tibble.
}
