% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_dirlist.R
\name{get_dirlist}
\alias{get_dirlist}
\title{Download file list from Entrez GEO repository ftp directory}
\usage{
get_dirlist(Accession)
}
\arguments{
\item{Accession}{GEO Accession, a charcter string.}
}
\description{
Downloads list of supplementary or MINiML formatted family file names from Entrez GEO repository.
}
\examples{
resp <- get_dirlist("GSE100206")
resp

}
