% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrez_docsums.R
\name{get_docsums}
\alias{get_docsums}
\title{Get entrez document summaries for UIDs}
\usage{
get_docsums(uid, db, chunksize = 100, ...)
}
\arguments{
\item{uid}{Character vector of UIDs.}

\item{db}{Entrez database, defaults to "gds" == GEO.}

\item{chunksize}{number of uids simultaneously queried, integer. Longer vectors will be split into smaller chunks.}

\item{...}{Further arguments to esummary API.}
}
\value{
A list of document summaries of class "xml_document" "xml_node".
}
\description{
Returns entrez document summaries for UIDs using esummary API.
}
