% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrez_docsums.R
\name{get_ids}
\alias{get_ids}
\title{Get UID vector}
\usage{
get_ids(query, db, retmax = 500, ...)
}
\arguments{
\item{query}{A GEO query string}

\item{db}{Entrez database. For example "gds" == GEO.}

\item{retmax}{Maximum number of records to return, default is 500.}

\item{...}{Further arguments to esearch API.}
}
\value{
A vector of UIDs.
}
\description{
Runs entrez search using esearch API and returns UIDs.
}
\examples{
\dontrun{
query <- 'expression profiling by high throughput sequencing[DataSet Type]'
ids <- get_ids(query, db = 'gds', retmax = 10)
}
}
