% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrez_docsums.R
\name{get_qsums}
\alias{get_qsums}
\title{Run GET request on Entrez database with UIDs.}
\usage{
get_qsums(uid, db, ...)
}
\arguments{
\item{db}{Entrez database, defaults to "gds" == GEO.}

\item{...}{Further arguments to esummary API.}

\item{ids}{Character vector of UIDs.}
}
