% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_dirlist.R
\name{munge_dirlist}
\alias{munge_dirlist}
\title{Returns supplementary file names from Entrez GEO}
\usage{
munge_dirlist(x)
}
\arguments{
\item{x}{a character vector of files at Entrez GEO ftp site.}
}
\value{
a data frame with columns date, size and suppfile.
Date is the supplementary file date, size is filesize in bytes.
Suppfile is the supplementary file name.
}
\description{
Replaces GEOquery package getDirlisting function, seems to be faster.
}
\examples{
\dontrun{

## Download supplementary file name
url <-  "ftp://ftp.ncbi.nlm.nih.gov/geo/series/GSE81nnn/GSE81555/suppl/"
r <- httr::GET(url)
cont <- httr::content(r, as = "text", encoding = "UTF-8")
suppl_filename <- munge_dirlist(cont)

## Download series matrix file name
url <-  "ftp://ftp.ncbi.nlm.nih.gov/geo/series/GSE81nnn/GSE81555/miniml/"
r <- httr::GET(url)
cont <- httr::content(r, as = "text", encoding = "UTF-8")
miniml_filename <- munge_dirlist(cont)

}

}
