% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_dirlist.R
\name{this_year}
\alias{this_year}
\title{Return current year when clock time is provided from entrez GEO ftp.}
\usage{
this_year(x)
}
\arguments{
\item{x}{character string of clock time e.g. '6:21' or year.}
}
\value{
character string with current year.
}
\description{
Entrez database shows clock time for files submitted on the current year.
}
