% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrez_docsums.R
\name{get_qsums}
\alias{get_qsums}
\title{Run GET request on Entrez database with UIDs.}
\usage{
get_qsums(uid, db, ...)
}
\arguments{
\item{uid}{Character vector of UIDs.}

\item{db}{Entrez database, defaults to "gds" == GEO.}

\item{...}{Further arguments to esummary API.}
}
\description{
Runs entrez query using esummary API, returns html response.
}
